various new files towards next-toilet and multi-language description tests, and some modifications to parking-spaces.spec.ts
